+++ src/main/resources/fxml/main.fxml
@@ -164,7 +164,6 @@
          </font>
       </Button>
       <Button fx:id="settingsBtn" layoutX="14.0" layoutY="14.0" mnemonicParsing="false" onAction="#onSettingsBtnAction" text="Settings" />
-      <Button fx:id="languageBtn" layoutX="78.0" layoutY="15.0" mnemonicParsing="false" onAction="#onLanguageBtnAction" text="EN" />
       <TextField fx:id="tempSetField" editable="true" layoutX="494.0" layoutY="343.0" prefHeight="26.0" prefWidth="83.0" />
       <Text fx:id="tempSetText" fill="white" layoutX="405.0" layoutY="362.0" strokeType="OUTSIDE" strokeWidth="0.0" text="t°з">
          <font>
