+++ src/main/resources/lang_en.properties
@@ -2,7 +2,6 @@
 
 # Main
 title = ITPC Advanced
-button.set.lang = EN
 button.settings = Settings
 button.get.report = Get Report
 button.scan = Scan
@@ -28,7 +27,9 @@ settings.shuffle.values = Shuffle values:
 settings.demo.mode = Demo mode:
 settings.connection.timeout = Connection timeout, ms:
 settings.max.deviation = Max deviation value, T°C:
-settings.button.save = Save settings
+settings.language = Language: 
+settings.button.set.lang = English
+settings.button.save = Close
 
 # DataFile Properties
 df.tc.type = Type:
