+++ src/main/resources/lang_ru.properties
@@ -1,7 +1,6 @@
 # Russian
 
 title = ITPC Advanced
-button.set.lang = RU
 button.settings = \u041D\u0430\u0441\u0442\u0440\u043E\u0439\u043A\u0438
 button.get.report = \u0421\u043A\u043E\u043F\u0438\u0440\u043E\u0432\u0430\u0442\u044C \u0440\u0435\u0437\u0443\u043B\u044C\u0442\u0430\u0442\u044B
 button.scan = \u0421\u043A\u0430\u043D\u0438\u0440\u043E\u0432\u0430\u0442\u044C
@@ -27,12 +26,14 @@ settings.shuffle.values = \u041F\u0435\u0440\u0435\u043C\u0435\u0448\u0438\u0432
 settings.demo.mode = \u0414\u0435\u043C\u043E \u0440\u0435\u0436\u0438\u043C:
 settings.connection.timeout = \u0412\u0440\u0435\u043C\u044F \u043D\u0430 \u043F\u043E\u043F\u044B\u0442\u043A\u0443 \u0441\u043E\u0435\u0434\u0438\u043D\u0435\u043D\u0438\u044F, \u043C\u0441:
 settings.max.deviation = \u041C\u0430\u043A\u0441. \u0434\u043E\u043F\u0443\u0441\u0442\u0438\u043C\u043E\u0435 \u043E\u0442\u043A\u043B\u043E\u043D\u0435\u043D\u0438\u0435, C: 
-settings.button.save = \u0421\u043E\u0445\u0440\u0430\u043D\u0438\u0442\u044C \u043D\u0430\u0441\u0442\u0440\u043E\u0439\u043A\u0438
+settings.language = \u042F\u0437\u044B\u043A: 
+settings.button.set.lang = Russian
+settings.button.save = \u0417\u0430\u043A\u0440\u044B\u0442\u044C
 
 # DataFile Properties
 df.tc.type = \u0422\u0438\u043F:  
-df.timestamp = \u0428\u0430\u0433, \u0441:  
-df.timestep = \u0412\u0440\u0435\u043C\u044F:  
+df.timestamp = \u0412\u0440\u0435\u043C\u044F:   
+df.timestep = \u0428\u0430\u0433, \u0441:  
 df.points.count = \u0422\u043E\u0447\u0435\u043A:  
 
 # ThermoCouple Types
